Organize MIFARE protocol partials by family
