+++ Elatec.NET.cs
@@ -35,7 +35,7 @@ namespace Elatec.NET
     ///     - e.g. <see cref="GetSingleChipAsync(bool)"/> providing detailed information</item>
     ///     </list>
     /// </summary>
-    public class TWN4ReaderDevice : IDisposable
+    public partial class TWN4ReaderDevice : IDisposable
     {
         private bool _disposed;
 
@@ -73,2348 +73,6 @@ namespace Elatec.NET
         /// 
         /// </summary>
         public static readonly int TIMEOUT = 20;
-
-        #region Low level APIs
-
-        #region API_SYS / System Functions
-
-        public static readonly byte API_SYS = 0;
-
-        // Not supported: SYSFUNC(API_SYS, 0, bool SysCall(TEnvSysCall* Env))
-
-        /// <summary>
-        /// This function is performing a reset of the firmware, which also includes a restart of the currently running App.
-        /// </summary>
-        /// <returns></returns>
-        public async Task ResetAsync()
-        {
-            await CallFunctionAsync(new byte[] { API_SYS, 1 });
-        }
-
-        /// <summary>
-        /// This function is performing a manual call of the boot loader. As a consequence the execution of the App is stopped.
-        /// </summary>
-        /// <returns></returns>
-        public async Task StartBootloaderAsync()
-        {
-            await CallFunctionAsync(new byte[] { API_SYS, 2 });
-        }
-
-        /// <summary>
-        /// Retrieve number of system ticks, specified in multiple of 1 milliseconds, since startup of the firmware.
-        /// </summary>
-        /// <returns></returns>
-        public async Task<uint> GetSysTicksAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_SYS, 3 });
-            uint ticks = parser.ParseUInt32();
-            return ticks;
-        }
-
-        /// <summary>
-        /// Retrieve version information.
-        /// </summary>
-        /// <returns></returns>
-        public async Task<string> GetVersionStringAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_SYS, 4, /* maxLen */ byte.MaxValue });
-            string version = parser.ParseAsciiString();
-            var subVersion = version.Split('/');
-            IsTWN4LegicReader = subVersion.Length >= 3 && subVersion[2].Contains('B');
-            return version;
-        }
-
-        /// <summary>
-        ///     Retrieve type of USB communication. This could by keyboard emulation or CDC emulation or some other
-        ///     value for future or custom implementations.
-        /// </summary>
-        /// <returns></returns>
-        public async Task<UsbType> GetUsbTypeAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_SYS, 5 });
-            var type = (UsbType)parser.ParseByte();
-            return type;
-        }
-
-        /// <summary>
-        /// Retrieve type of underlying TWN4 hardware.
-        /// </summary>
-        /// <returns></returns>
-        public async Task<DeviceType> GetDeviceTypeAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_SYS, 6 });
-            var type = (DeviceType)parser.ParseByte();
-            return type;
-        }
-
-        /// <summary>
-        ///     The device enters the sleep state for a specified time. During sleep state, the device reduces the current
-        ///     consumption to a value, which depends on the mode of sleep.
-        /// </summary>
-        /// <param name="ticks">Time, specified in milliseconds, the device should enter the sleep state.</param>
-        /// <param name="flags">See TWN4 API Reference.</param>
-        /// <returns>See TWN4 API Reference.</returns>
-        public async Task<byte> SleepAsync(uint ticks, uint flags)
-        {
-            List<byte> bytes = new List<byte> { API_RF, 7 };
-            bytes.AddUInt32(ticks);
-            bytes.AddUInt32(flags);
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var result = parser.ParseByte();
-            return result;
-        }
-
-        /// <summary>
-        /// This function returns a UID, which is unique to the specific TWN4 device.
-        /// </summary>
-        /// <returns></returns>
-        public async Task<byte[]> GetDeviceUidAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_SYS, 8 });
-            byte[] result = parser.ParseFixByteArray(12);
-            return result;
-        }
-
-        /// <summary>
-        ///     This function allows to set parameters, which influence the behaviour of the TWN4 firmware. See also
-        ///     chapter System Parameters of TWN4 API Reference for a description of the TLV list and all available parameters.
-        /// </summary>
-        /// <param name="TLV">The raw bytes of the TLV list. Do not include TLV_END, as it is appended automatically!</param>
-        /// <returns>The function returns true, if the parameters were set to the new value. Otherwise
-        ///     the function returns false.</returns>
-        /// <remarks>SYSFUNC(API_SYS, 9, bool SetParameters(const byte* TLV,int ByteCount))</remarks>
-        public async Task<bool> SetParametersAsync(byte[] TLV)
-        {
-            List<byte> bytes = new List<byte> { API_SYS, 9 };
-            bytes.Add((byte)(TLV.Length + 1));
-            bytes.AddRange(TLV);
-            bytes.Add(0); // TLV_END
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var result = parser.ParseBool();
-            return result;
-        }
-
-        /// <summary>
-        /// This function is used to retrieve internal system errors of the reader. Do not deduce protocol or communication errors from this function call.
-        /// </summary>
-        /// <returns></returns>
-        /// <remarks>SYSFUNC(API_SYS,10, unsigned int GetLastError(void))</remarks>
-        public async Task<ReaderError> GetLastErrorAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_SYS, 10 });
-            var result = (ReaderError)parser.ParseUInt32();
-            return result;
-        }
-
-        // Not supported: SYSFUNC(API_SYS,11, int Diagnostic(int Mode,const void* In,int InLen,void* Out,int* OutLen,int MaxOutLen))
-
-        /// <summary>
-        /// Get the product serial number of the TWN device.
-        /// </summary>
-        /// <returns></returns>
-        public async Task<string> GetProdSerNoAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_SYS, 13, /* maxBytes */ byte.MaxValue });
-            string result = parser.ParseAsciiString();
-            return result;
-        }
-
-        // Not supported: SYSFUNC(API_SYS,14, bool SetInterruptHandler(TInterruptHandler InterruptHandler, int IntNo))
-
-        /// <summary>
-        /// Retrieve version information.
-        /// </summary>
-        /// <returns></returns>
-        /// <remarks>SYSFUNC(API_SYS,15, void GetVersionInfo(TVersionInfo* VersionInfo)).<br/>
-        ///     This internal method is not documented in TWN4 API reference.
-        /// </remarks>
-        public async Task<VersionInfo> GetVersionInfoAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_SYS, 15 });
-            var info = new VersionInfo();
-            info.Compatibility = parser.ParseUInt16();
-            info.BootBranch = parser.ParseUInt16();
-            var minor = parser.ParseByte();
-            var major = parser.ParseByte();
-            info.BootVersion = new Version(major, minor);
-            info.FirmwareKeyType = parser.ParseUInt16();
-            info.BranchNum = parser.ParseByte();
-            info.BranchChar = (char)parser.ParseByte();
-            minor = parser.ParseByte();
-            major = parser.ParseByte();
-            info.FirmwareVersion = new Version(major, minor);
-            info.AppChars = parser.ParseFixByteArray(4);
-            minor = parser.ParseByte();
-            major = parser.ParseByte();
-            info.AppVersion = new Version(major, minor);
-
-            return info;
-        }
-
-        public class VersionInfo
-        {
-            public int Compatibility { get; set; }
-            public int BootBranch { get; set; }
-            public Version BootVersion { get; set; }
-            public int FirmwareKeyType { get; set; }
-            public byte BranchNum { get; set; }
-            /// <summary>
-            /// 'K' = Keyboard, 'C' = CDC
-            /// </summary>
-            public char BranchChar { get; set; }
-            public Version FirmwareVersion { get; set; }
-            /// <summary>
-            /// e.g. "STD", "STDC", "PRS" = Simple Protocol
-            /// </summary>
-            public byte[] AppChars { get; set; }
-            public Version AppVersion { get; set; }
-        }
-
-        // Not supported: SYSFUNC(API_SYS,16, bool ReadInfoValue(int Index, int FilterType, int* Type, int* Length, byte* Value, int MaxLength))
-        // Not supported: SYSFUNC(API_SYS,17, bool WriteInfoValue(int Type, int Length,const byte* Value))
-        // Not supported: SYSFUNC(API_SYS,18, bool GetCustomKeyID(byte* CustomKeyID, int* Length, int MaxLength))
-        // Not supported: SYSFUNC(API_SYS,19, bool GetParameters(const byte* Types,int TypeCount,byte* TLVBytes,int* TLVByteCount,int TLVMaxByteCount))
-
-        #endregion
-
-        #region API_PERIPH / Periphery Functions
-
-        public static readonly byte API_PERIPH = 4;
-
-        /// <summary>
-        ///     Use this function to configure one or several GPIOs as output. Each output can be configured to have an
-        ///     integrated pull-up or pull-down resistor.The output driver characteristic is either Push-Pull or Open Drain.
-        /// </summary>
-        /// <param name="bits">Specify the GPIOs that shall be configured for output. Several GPIOs can
-        ///     be configured simultaneously by using the bitwise or-operator (|).</param>
-        /// <param name="pullUpDown">Specify the behaviour of the internal weak pull-up/down resistor.</param>
-        /// <param name="outputType">Specify the output driver characteristic.</param>
-        /// <returns></returns>
-        public async Task GpioConfigureOutputsAsync(Gpios bits, GpioPullType pullUpDown, GpioOutputType outputType)
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 0, (byte)bits, (byte)pullUpDown, (byte)outputType });
-        }
-
-        /// <summary>
-        ///     Use this function to configure one or several GPIOs as input. Each output can be configured to have an
-        ///     integrated pull-up or pull-down resistor, alternatively it can be left floating.
-        /// </summary>
-        /// <param name="bits">Specify the GPIOs that shall be configured for input. Several GPIOs can
-        ///     be configured simultaneously by using the bitwise or-operator (|).</param>
-        /// <param name="pullUpDown">Specify the behaviour of the internal weak pull-up/down resistor.</param>
-        /// <returns></returns>
-        public async Task GpioConfigureInputsAsync(Gpios bits, GpioPullType pullUpDown)
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 1, (byte)bits, (byte)pullUpDown });
-        }
-
-        /// <summary>
-        ///     Use this function to set one or several GPIOs to logical high level.
-        ///     The respective ports must have been configured to output in advance.
-        /// </summary>
-        /// <param name="bits">Specify the GPIOs that shall be set to a logical level. Several GPIOs can
-        ///     be handled simultaneously by using the bitwise or-operator (|).</param>
-        /// <returns></returns>
-        public async Task GpioSetBitsAsync(Gpios bits)
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 2, (byte)bits });
-        }
-
-        /// <summary>
-        ///     Use this function to set one or several GPIOs to logical low level.
-        ///     The respective ports must have been configured to output in advance.
-        /// </summary>
-        /// <param name="bits">Specify the GPIOs that shall be set to a logical level. Several GPIOs can
-        ///     be handled simultaneously by using the bitwise or-operator (|).</param>
-        /// <returns></returns>
-        public async Task GpioClearBitsAsync(Gpios bits)
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 3, (byte)bits });
-        }
-
-        /// <summary>
-        ///     Use this function to toggle the logical level of one or several GPIOs.
-        ///     The respective ports must have been configured to output in advance.
-        /// </summary>
-        /// <param name="bits">Specify the GPIOs that shall be set to a logical level. Several GPIOs can
-        ///     be handled simultaneously by using the bitwise or-operator (|).</param>
-        /// <returns></returns>
-        public async Task GpioToggleBitsAsync(Gpios bits)
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 4, (byte)bits });
-        }
-
-        /// <summary>
-        ///     Use this function to generate a pulse-width modulated square waveform with constant frequency on one
-        ///     or several GPIOs. The respective ports must have been configured to output in advance.
-        /// </summary>
-        /// <param name="bits">Specify the GPIOs that shall generate the waveform.</param>
-        /// <param name="timeHi">Specify the duration for logical high level in milliseconds.</param>
-        /// <param name="timeLo">Specify the duration for logical low level in milliseconds.</param>
-        /// <returns></returns>
-        public async Task GpioBlinkBitsAsync(Gpios bits, ushort timeHi, ushort timeLo)
-        {
-            List<byte> bytes = new List<byte> { API_PERIPH, 5 };
-            bytes.Add((byte)bits);
-            bytes.AddUInt16(timeHi);
-            bytes.AddUInt16(timeLo);
-            await CallFunctionAsync(bytes.ToArray());
-        }
-
-        /// <summary>
-        /// Use this function to read the logical level of one GPIO that has been configured as input.
-        /// </summary>
-        /// <param name="bit">Specify the GPIO that shall be read.</param>
-        /// <returns>If the GPIO has logical high level, the return value is 1, otherwise it is 0.</returns>
-        public async Task<bool> GpioTestBitAsync(Gpios bit)
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_PERIPH, 6, (byte)bit });
-            var result = parser.ParseBool();
-            return result;
-        }
-
-        /// <summary>
-        /// Play a beep on the device.
-        /// </summary>
-        /// <param name="volume">Specify the volume in percent from 0 to 100.</param>
-        /// <param name="frequency">Specify the frequency in Hertz from 500 to 10000.</param>
-        /// <param name="onTime">Specify the duration of the beep in milliseconds from 0 to 10000000.</param>
-        /// <param name="offTime">Specify the length of the pause after the beep in milliseconds from 0 to
-        ///     10000000. This is useful for generating melodies.If this is not required, the
-        ///     parameter may have the value 0.</param>
-        /// <returns></returns>
-        public async Task BeepAsync(byte volume, ushort frequency, ushort onTime, ushort offTime)
-        {
-            List<byte> bytes = new List<byte> { API_PERIPH, 7 };
-            bytes.Add(volume);
-            bytes.AddUInt16(frequency);
-            bytes.AddUInt16(onTime);
-            bytes.AddUInt16(offTime);
-            await CallFunctionAsync(bytes.ToArray());
-        }
-
-        public async Task DiagLedOnAsync()
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 8 });
-        }
-
-        public async Task DiagLedOffAsync()
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 9 });
-        }
-
-        public async Task DiagLedToggleAsync()
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 10 });
-        }
-
-        public async Task<bool> DiagLedIsOnAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_PERIPH, 11 });
-            var result = parser.ParseBool();
-            return result;
-        }
-
-        // TODO: SYSFUNC(API_PERIPH,12, void SendWiegand(int GPIOData0, int GPIOData1, int PulseTime, int IntervalTime,const byte* Bits,int BitCount))
-        // TODO: SYSFUNC(API_PERIPH,13, void SendOmron(int GPIOClock, int GPIOData, int T1, int T2, int T3,const byte* Bits,int BitCount))
-        // Not supported: SYSFUNC(API_PERIPH,14, bool GPIOPlaySequence(const int* NewSequence,int ByteCount))
-        // Not supported: SYSFUNC(API_PERIPH,15, void GPIOStopSequence(void))
-
-        /// <summary>
-        /// Use this function to initialize the respective GPIOs to drive LEDs.
-        /// </summary>
-        /// <param name="leds">Specify the GPIOs that shall be configured for LED operation.</param>
-        /// <returns></returns>
-        public async Task LedInitAsync()
-        {
-            await LedInitAsync(Leds.All);
-        }
-
-        /// <summary>
-        /// Use this function to initialize the respective GPIOs to drive LEDs.
-        /// </summary>
-        /// <param name="leds">Specify the GPIOs that shall be configured for LED operation.</param>
-        /// <returns></returns>
-        public async Task LedInitAsync(Leds leds)
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 16, (byte)leds });
-        }
-
-        /// <summary>
-        /// Use this function to set one or several LEDs on.
-        /// </summary>
-        /// <param name="leds">Specify the LEDs that shall be set on.</param>
-        /// <returns></returns>
-        public async Task LedOnAsync(Leds leds)
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 17, (byte)leds });
-        }
-
-        /// <summary>
-        /// Use this function to set one or several LEDs off.
-        /// </summary>
-        /// <param name="leds">Specify the LEDs that shall be set off.</param>
-        /// <returns></returns>
-        public async Task LedOffAsync(Leds leds)
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 18, (byte)leds });
-        }
-
-        /// <summary>
-        /// Use this function to toggle one or several LEDs.
-        /// </summary>
-        /// <param name="leds">Specify the LEDs that shall be toggled.</param>
-        /// <returns></returns>
-        public async Task LedToggleAsync(Leds leds)
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 19, (byte)leds });
-        }
-
-        /// <summary>
-        /// Use this function to let one or several LEDs blink.
-        /// </summary>
-        /// <param name="leds">Specify the LEDs that shall blink.</param>
-        /// <param name="onTime">Specify the on-time in milliseconds.</param>
-        /// <param name="offTime">Specify the off-time in milliseconds.</param>
-        /// <returns></returns>
-        public async Task LedBlinkAsync(Leds leds, ushort onTime, ushort offTime)
-        {
-            List<byte> bytes = new List<byte> { API_PERIPH, 20 };
-            bytes.Add((byte)leds);
-            bytes.AddUInt16(onTime);
-            bytes.AddUInt16(offTime);
-            await CallFunctionAsync(bytes.ToArray());
-        }
-
-        // Not supported: SYSFUNC(API_PERIPH,21,bool GPIOConfigureInterrupt(int GPIOBits,bool Enable,int Edge))
-
-        /// <summary>
-        /// Turn on beep with infinite length.
-        /// </summary>
-        /// <param name="volume">Specify the volume in percent from 0 to 100.</param>
-        /// <param name="frequency">Specify the frequency in Hertz from 500 to 10000.</param>
-        /// <returns></returns>
-        public async Task BeepOnAsync(byte volume, ushort frequency)
-        {
-            List<byte> bytes = new List<byte> { API_PERIPH, 22 };
-            bytes.Add(volume);
-            bytes.AddUInt16(frequency);
-            await CallFunctionAsync(bytes.ToArray());
-        }
-
-        /// <summary>
-        /// Turn off beep.
-        /// </summary>
-        /// <returns></returns>
-        public async Task BeepOffAsync()
-        {
-            await CallFunctionAsync(new byte[] { API_PERIPH, 23 });
-        }
-
-        // Not supported: SYSFUNC(API_PERIPH,24,void PlayMelody(const byte *Melody,int MelodyLength))
-
-        #endregion
-
-        #region API_RF
-
-        public static readonly byte API_RF = 5;
-
-        // API_RF function numbers (Simple Protocol: 0x0500..0x0504)
-        public static readonly byte RF_SEARCHTAG = 0;
-        public static readonly byte RF_SETRFOFF = 1;
-        public static readonly byte RF_SETTAGTYPES = 2;
-        public static readonly byte RF_GETTAGTYPES = 3;
-        public static readonly byte RF_GETSUPPORTEDTAGTYPES = 4;
-
-        /// <summary>
-        /// Calculates the TagTypes bitmask value for a given TagType code (see Appendix A in the Simple Protocol spec).
-        /// </summary>
-        /// <param name="tagType">The TagType code (e.g. 0x80 for MIFARE).</param>
-        /// <returns>The corresponding UInt32 mask bit (TAGMASK = 1 &lt;&lt; (TagType &amp; 0x1F)).</returns>
-        public static uint TagMaskFromTagType(byte tagType) => 1u << (tagType & 0x1F);
-
-
-        /// <summary>
-        ///     Use this function to search a transponder in the reading range of TWN4. TWN4 is searching for all types
-        ///     of transponders, which have been specified via function SetTagTypes. If a transponder has been found,
-        ///     tag type, length of ID and ID data itself are returned.
-        /// </summary>
-        /// <remarks>
-        ///     Simple Protocol command: 0x0500 (API_RF / SearchTag).
-        /// </remarks>
-        /// <returns>
-        ///     A <see cref="SearchTagResult" /> or <see langword="null" /> if no tag was detected.
-        ///     This overload requests the maximum ID length (MaxIDBytes = 0xFF).
-        /// </returns>
-        public async Task<SearchTagResult> SearchTagAsync()
-        {
-            return await SearchTagAsync(byte.MaxValue);
-        }
-
-        /// <summary>
-        ///     Use this function to search a transponder in the reading range of TWN4. TWN4 is searching for all types
-        ///     of transponders, which have been specified via function SetTagTypes. If a transponder has been found,
-        ///     tag type, length of ID and ID data itself are returned.
-        /// </summary>
-        /// <remarks>
-        ///     Simple Protocol command: 0x0500 (API_RF / SearchTag).
-        /// </remarks>
-        /// <param name="maxIdBytes">Maximum number of ID bytes to return (MaxIDBytes).</param>
-        /// <returns>A <see cref="SearchTagResult" /> or <see langword="null" /> if no tag was detected.</returns>
-        public async Task<SearchTagResult> SearchTagAsync(byte maxIdBytes)
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_RF, RF_SEARCHTAG, maxIdBytes });
-            var found = parser.ParseBool();
-            if (!found)
-                return null;
-
-            return new SearchTagResult
-            {
-                ChipType = (ChipType)parser.ParseByte(),
-                IDBitCount = parser.ParseByte(),
-                IDBytes = parser.ParseVarByteArray()
-            };
-        }
-
-        public class SearchTagResult
-        {
-            /// <summary>
-            /// Property is called TagType in the API.
-            /// </summary>
-            public ChipType ChipType { get; set; }
-            public byte IDBitCount { get; set; }
-            public byte[] IDBytes { get; set; }
-
-            public string IDHexString
-            {
-                get
-                {
-                    return ByteArrayConverter.GetStringFrom(IDBytes);
-                }
-            }
-        }
-
-        /// <summary>
-        ///     Turn off RF field. If no further operations are required on a transponder found via function SearchTag you
-        ///     may use this command to minimize power consumption of TWN4.
-        /// </summary>
-        /// <returns></returns>
-        public async Task SetRFOffAsync()
-        {
-            await CallFunctionAsync(new byte[] { API_RF, RF_SETRFOFF });
-        }
-
-        /// <summary>
-        /// Use this function to configure the transponders, which are searched by function SearchTag.
-        /// </summary>
-        /// <remarks>
-        /// The parameters are bitmasks (UInt32) of enabled tag technologies, split into LF and HF.
-        /// If you have "TagType codes" (e.g. 0x80 for MIFARE), convert them using <see cref="TagMaskFromTagType(byte)"/>.
-        /// </remarks>
-        /// <param name="lfTagTypes">LF tag type mask.</param>
-        /// <param name="hfTagTypes">HF tag type mask.</param>
-        public async Task SetTagTypesAsync(LFTagTypes lfTagTypes, HFTagTypes hfTagTypes)
-        {
-            await SetTagTypesAsync((uint)lfTagTypes, (uint)hfTagTypes);
-        }
-
-        /// <summary>
-        /// Use this function to configure the transponders, which are searched by function SearchTag.
-        /// </summary>
-        /// <remarks>
-        /// This overload takes the raw UInt32 masks exactly as defined by the Simple Protocol command 0x0502.
-        /// </remarks>
-        /// <param name="tagTypesLF">LF tag types bitmask.</param>
-        /// <param name="tagTypesHF">HF tag types bitmask.</param>
-        public async Task SetTagTypesAsync(uint tagTypesLF, uint tagTypesHF)
-        {
-            List<byte> bytes = new List<byte> { API_RF, RF_SETTAGTYPES };
-            bytes.AddUInt32(tagTypesLF);
-            bytes.AddUInt32(tagTypesHF);
-            await CallFunctionAsync(bytes.ToArray());
-        }
-
-        /// <summary>
-        ///     This function returns the transponder types currently being searched for by function SearchTag separated
-        ///     by frequency (LF and HF).
-        /// </summary>
-        /// <returns>Tag types.</returns>
-        public async Task<GetTagTypesResult> GetTagTypesAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_RF, RF_GETTAGTYPES });
-            var lf = parser.ParseUInt32();
-            var hf = parser.ParseUInt32();
-
-            return new GetTagTypesResult() { LFTagTypes = (LFTagTypes)lf, HFTagTypes = (HFTagTypes)hf };
-        }
-
-        /// <summary>
-        /// Tag type masks currently configured for <see cref="SearchTagAsync()"/> (Simple Protocol: 0x0503).
-        /// </summary>
-        public class GetTagTypesResult
-        {
-            /// <summary>LF tag types bitmask.</summary>
-            public LFTagTypes LFTagTypes { get; internal set; }
-            /// <summary>HF tag types bitmask.</summary>
-            public HFTagTypes HFTagTypes { get; internal set; }
-        }
-
-
-        /// <summary>
-        ///     This function returns the transponder types, which are actually supported by the individual TWN4 separated
-        ///     by frequency (LF and HF). Also the P-option is taken into account. This means, if the specific TWN4
-        ///     has no option P, the appropriate transponders are not returned as supported type of transponder.
-        /// </summary>
-        /// <returns>Tag types.</returns>
-        public async Task<GetSupportedTagTypesResult> GetSupportedTagTypesAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_RF, RF_GETSUPPORTEDTAGTYPES });
-            var lf = parser.ParseUInt32();
-            var hf = parser.ParseUInt32();
-
-            return new GetSupportedTagTypesResult() { LFTagTypes = (LFTagTypes)lf, HFTagTypes = (HFTagTypes)hf };
-        }
-
-        /// <summary>
-        /// Tag type masks supported by this device (Simple Protocol: 0x0504).
-        /// </summary>
-        public class GetSupportedTagTypesResult
-        {
-            /// <summary>LF tag types bitmask.</summary>
-            public LFTagTypes LFTagTypes { get; internal set; }
-            /// <summary>HF tag types bitmask.</summary>
-            public HFTagTypes HFTagTypes { get; internal set; }
-        }
-
-        #endregion
-
-        #region API_MIFARECLASSIC / Mifare Classic Functions
-
-        public static readonly byte API_MIFARECLASSIC = 11;
-
-        public static readonly byte MIFARE_CLASSIC_LOGIN = 0;
-        public static readonly byte MIFARE_CLASSIC_READBLOCK = 1;
-        public static readonly byte MIFARE_CLASSIC_WRITEBLOCK = 2;
-
-
-        public static readonly byte MIFARE_CLASSIC_READVALUEBLOCK = 3;
-        public static readonly byte MIFARE_CLASSIC_WRITEVALUEBLOCK = 4;
-        public static readonly byte MIFARE_CLASSIC_INCREMENTVALUEBLOCK = 5;
-        public static readonly byte MIFARE_CLASSIC_DECREMENTVALUEBLOCK = 6;
-        public static readonly byte MIFARE_CLASSIC_COPYVALUEBLOCK = 7;
-        /// <summary>
-        /// Login to a Mifare Classic single Sector.
-        /// </summary>
-        /// <param name="key"></param>
-        /// <param name="keyType"></param>
-        /// <param name="sectorNumber"></param>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareClassic_LoginAsync(string key, byte keyType, byte sectorNumber)
-        {
-            List<byte> bytes = new List<byte> { API_MIFARECLASSIC, MIFARE_CLASSIC_LOGIN };
-            bytes.AddRange(ByteArrayConverter.GetBytesFrom(key));
-            bytes.Add(keyType);
-            bytes.Add(sectorNumber);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Read Data from Classic Chip
-        /// </summary>
-        /// <param name="blockNumber">DataBlock Number</param>
-        /// <returns>DATA</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<byte[]> MifareClassic_ReadBlockAsync(byte blockNumber)
-        {
-            List<byte> bytes = new List<byte> { API_MIFARECLASSIC, MIFARE_CLASSIC_READBLOCK };
-            bytes.Add(blockNumber);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                return parser.ParseFixByteArray(16);
-            }
-            else
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Write Data to Classic Chip
-        /// </summary>
-        /// <param name="data">16 Bytes of Data to Write</param>
-        /// <param name="blockNumber">DataBlock Number</param>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareClassic_WriteBlockAsync(byte[] data, byte blockNumber)
-        {
-            List<byte> bytes = new List<byte> { API_MIFARECLASSIC, MIFARE_CLASSIC_WRITEBLOCK, blockNumber };
-            bytes.AddRange(data);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-/// <summary>
-/// Read a Value Block from a MIFARE Classic transponder.
-/// </summary>
-/// <param name="blockNumber">Value block number.</param>
-/// <returns>Value stored in the block (32-bit, returned as signed int).</returns>
-/// <exception cref="ReaderException"></exception>
-public async Task<int> MifareClassic_ReadValueBlockAsync(byte blockNumber)
-{
-    List<byte> bytes = new List<byte> { API_MIFARECLASSIC, MIFARE_CLASSIC_READVALUEBLOCK };
-    bytes.Add(blockNumber);
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (success)
-    {
-        var value = parser.ParseUInt32();
-        return unchecked((int)value);
-    }
-
-    throw new ReaderException("Call was not successfull, error " +
-                              Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-}
-
-/// <summary>
-/// Write a Value Block to a MIFARE Classic transponder.
-/// </summary>
-/// <param name="blockNumber">Value block number.</param>
-/// <param name="value">Value to write (32-bit).</param>
-/// <exception cref="ReaderException"></exception>
-public async Task MifareClassic_WriteValueBlockAsync(byte blockNumber, int value)
-{
-    List<byte> bytes = new List<byte> { API_MIFARECLASSIC, MIFARE_CLASSIC_WRITEVALUEBLOCK };
-    bytes.Add(blockNumber);
-    bytes.AddUInt32(unchecked((uint)value));
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " +
-                                  Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-
-/// <summary>
-/// Increment a Value Block on a MIFARE Classic transponder.
-/// </summary>
-/// <param name="blockNumber">Value block number.</param>
-/// <param name="value">Increment amount (32-bit).</param>
-/// <exception cref="ReaderException"></exception>
-public async Task MifareClassic_IncrementValueBlockAsync(byte blockNumber, int value)
-{
-    List<byte> bytes = new List<byte> { API_MIFARECLASSIC, MIFARE_CLASSIC_INCREMENTVALUEBLOCK };
-    bytes.Add(blockNumber);
-    bytes.AddUInt32(unchecked((uint)value));
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " +
-                                  Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-
-/// <summary>
-/// Decrement a Value Block on a MIFARE Classic transponder.
-/// </summary>
-/// <param name="blockNumber">Value block number.</param>
-/// <param name="value">Decrement amount (32-bit).</param>
-/// <exception cref="ReaderException"></exception>
-public async Task MifareClassic_DecrementValueBlockAsync(byte blockNumber, int value)
-{
-    List<byte> bytes = new List<byte> { API_MIFARECLASSIC, MIFARE_CLASSIC_DECREMENTVALUEBLOCK };
-    bytes.Add(blockNumber);
-    bytes.AddUInt32(unchecked((uint)value));
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " +
-                                  Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-
-/// <summary>
-/// Copy a Value Block on a MIFARE Classic transponder.
-/// </summary>
-/// <param name="sourceBlockNumber">Source value block.</param>
-/// <param name="destinationBlockNumber">Destination value block.</param>
-/// <exception cref="ReaderException"></exception>
-public async Task MifareClassic_CopyValueBlockAsync(byte sourceBlockNumber, byte destinationBlockNumber)
-{
-    List<byte> bytes = new List<byte> { API_MIFARECLASSIC, MIFARE_CLASSIC_COPYVALUEBLOCK };
-    bytes.Add(sourceBlockNumber);
-    bytes.Add(destinationBlockNumber);
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " +
-                                  Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-        #endregion
-
-        #region API_MIFAREULTRALIGHT / Mifare Ultralight Functions
-
-        public static readonly byte API_MIFAREULTRALIGHT = 12;
-
-        /// <summary>
-        ///     Though the page size of this transponder family is 4 bytes, the transponder always returns 16 bytes of data.
-        ///     This is achieved by reading four consecutive data pages, e.g. if page 4 is to be read, the transponder also
-        ///     returns the content of page 5, 6 and 7. The transponder incorporates an integrated roll-back mechanism
-        ///     if reading is done beyond the last physical available page address.E.g., in case of reading page 14 of
-        ///     MIFARE Ultralight this would result in reading page 14, 15, 0, 1.
-        /// </summary>
-        /// <param name="page">Specify the address of the page to be read. The valid range of this parameter
-        ///     is between 0 and 15 (Ultralight) or 0 and 43 (Ultralight C).</param>
-        /// <returns></returns>
-        public async Task<byte[]> MifareUltralight_ReadPageAsync(byte page)
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_MIFAREULTRALIGHT, 0, page });
-            var success = parser.ParseBool();
-            if (success)
-            {
-                var result = parser.ParseFixByteArray(16);
-                return result;
-            }
-            return null;
-        }
-
-
-/// <summary>
-/// Write 4 bytes to a MIFARE Ultralight (and Ultralight C) page.
-/// </summary>
-/// <param name="page">Page address to write.</param>
-/// <param name="data">Exactly 4 bytes of data.</param>
-public async Task MifareUltralight_WritePageAsync(byte page, byte[] data)
-{
-    if (data == null) throw new ArgumentNullException(nameof(data));
-    if (data.Length != 4) throw new ArgumentException("MIFARE Ultralight page size is 4 bytes.", nameof(data));
-
-    List<byte> bytes = new List<byte> { API_MIFAREULTRALIGHT, 1, page };
-    bytes.AddRange(data);
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-
-/// <summary>
-/// Authenticate on MIFARE Ultralight C using a 16 byte key.
-/// </summary>
-/// <param name="keyHex">16 byte key as hex string (32 hex chars).</param>
-public async Task MifareUltralightC_AuthenticateAsync(string keyHex)
-{
-    if (keyHex == null) throw new ArgumentNullException(nameof(keyHex));
-
-    var key = ByteArrayConverter.GetBytesFrom(keyHex);
-    if (key == null || key.Length != 16)
-        throw new ArgumentException("Ultralight C key must be 16 bytes (32 hex characters).", nameof(keyHex));
-
-    await MifareUltralightC_AuthenticateAsync(key);
-}
-
-/// <summary>
-/// Authenticate on MIFARE Ultralight C using a 16 byte key.
-/// </summary>
-/// <param name="key">16 byte key.</param>
-public async Task MifareUltralightC_AuthenticateAsync(byte[] key)
-{
-    if (key == null) throw new ArgumentNullException(nameof(key));
-    if (key.Length != 16) throw new ArgumentException("Ultralight C key must be 16 bytes.", nameof(key));
-
-    List<byte> bytes = new List<byte> { API_MIFAREULTRALIGHT, 2 };
-    bytes.AddRange(key);
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-
-/// <summary>
-/// Authenticate on MIFARE Ultralight C using a key stored in the SAM (optional diversification input).
-/// </summary>
-/// <param name="keyNo">SAM key number.</param>
-/// <param name="keyVersion">SAM key version.</param>
-/// <param name="divInput">Diversification input (may be null/empty).</param>
-public async Task MifareUltralightC_SAMAuthenticateAsync(byte keyNo, byte keyVersion, byte[] divInput = null)
-{
-    var div = divInput ?? Array.Empty<byte>();
-    if (div.Length > byte.MaxValue) throw new ArgumentException("DIVInput too large (max 255 bytes).", nameof(divInput));
-
-    List<byte> bytes = new List<byte> { API_MIFAREULTRALIGHT, 3, keyNo, keyVersion, (byte)div.Length };
-    bytes.AddRange(div);
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-
-/// <summary>
-/// Write Ultralight C key to tag from SAM (optional diversification input).
-/// </summary>
-/// <param name="keyNo">SAM key number.</param>
-/// <param name="keyVersion">SAM key version.</param>
-/// <param name="divInput">Diversification input (may be null/empty).</param>
-public async Task MifareUltralightC_WriteKeyFromSAMAsync(byte keyNo, byte keyVersion, byte[] divInput = null)
-{
-    var div = divInput ?? Array.Empty<byte>();
-    if (div.Length > byte.MaxValue) throw new ArgumentException("DIVInput too large (max 255 bytes).", nameof(divInput));
-
-    List<byte> bytes = new List<byte> { API_MIFAREULTRALIGHT, 4, keyNo, keyVersion, (byte)div.Length };
-    bytes.AddRange(div);
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-
-        /// <summary>
-        /// The Fast Read function reads a number of pages beginning at a starting page from the transponder.
-        /// </summary>
-        /// <param name="startPage">Specify the address of the starting page.</param>
-        /// <param name="numberOfPages">Specify the number of pages to be read.</param>
-        /// <returns></returns>
-        public async Task<byte[]> MifareUltralightEV1_FastReadAsync(byte startPage, byte numberOfPages)
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_MIFAREULTRALIGHT, 5, startPage, numberOfPages });
-            var success = parser.ParseBool();
-            if (success)
-            {
-                var result = parser.ParseVarByteArray();
-                return result;
-            }
-            return null;
-        }
-
-
-/// <summary>
-/// Increment one of the Ultralight EV1 counters.
-/// </summary>
-/// <param name="counterAddr">Counter address.</param>
-/// <param name="incrValue">Increment value (UInt32).</param>
-public async Task MifareUltralightEV1_IncCounterAsync(byte counterAddr, uint incrValue)
-{
-    List<byte> bytes = new List<byte> { API_MIFAREULTRALIGHT, 6, counterAddr };
-    bytes.AddUInt32(incrValue);
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-
-/// <summary>
-/// Read one of the Ultralight EV1 counters.
-/// </summary>
-/// <param name="counterAddr">Counter address.</param>
-/// <returns>Counter value or null on failure.</returns>
-public async Task<uint?> MifareUltralightEV1_ReadCounterAsync(byte counterAddr)
-{
-    var parser = await CallFunctionAsync(new byte[] { API_MIFAREULTRALIGHT, 7, counterAddr });
-    var success = parser.ParseBool();
-    if (success)
-    {
-        return parser.ParseUInt32();
-    }
-    return null;
-}
-
-/// <summary>
-/// Read the ECC signature from an Ultralight EV1 tag.
-/// </summary>
-/// <returns>32 byte ECC signature or null on failure.</returns>
-public async Task<byte[]> MifareUltralightEV1_ReadSigAsync()
-{
-    var parser = await CallFunctionAsync(new byte[] { API_MIFAREULTRALIGHT, 8 });
-    var success = parser.ParseBool();
-    if (success)
-    {
-        return parser.ParseFixByteArray(32);
-    }
-    return null;
-}
-
-/// <summary>
-/// Get version information from an Ultralight EV1 tag.
-/// </summary>
-/// <returns>8 byte version information or null on failure.</returns>
-public async Task<byte[]> MifareUltralightEV1_GetVersionAsync()
-{
-    var parser = await CallFunctionAsync(new byte[] { API_MIFAREULTRALIGHT, 9 });
-    var success = parser.ParseBool();
-    if (success)
-    {
-        return parser.ParseFixByteArray(8);
-    }
-    return null;
-}
-
-/// <summary>
-/// Perform password authentication (PWD_AUTH) on Ultralight EV1 / NTAG compatible tags.
-/// </summary>
-/// <param name="password">4 byte password.</param>
-/// <param name="pwdAck">2 byte PACK / PwdAck.</param>
-public async Task MifareUltralightEV1_PwdAuthAsync(byte[] password, byte[] pwdAck)
-{
-    if (password == null) throw new ArgumentNullException(nameof(password));
-    if (pwdAck == null) throw new ArgumentNullException(nameof(pwdAck));
-    if (password.Length != 4) throw new ArgumentException("Password must be 4 bytes.", nameof(password));
-    if (pwdAck.Length != 2) throw new ArgumentException("PwdAck must be 2 bytes.", nameof(pwdAck));
-
-    List<byte> bytes = new List<byte> { API_MIFAREULTRALIGHT, 10 };
-    bytes.AddRange(password);
-    bytes.AddRange(pwdAck);
-
-    var parser = await CallFunctionAsync(bytes.ToArray());
-    var success = parser.ParseBool();
-
-    if (!success)
-    {
-        throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-    }
-}
-
-/// <summary>
-/// Check tearing event flag for an Ultralight EV1 counter.
-/// </summary>
-/// <param name="counterAddr">Counter address.</param>
-/// <returns>ValidFlag byte or null on failure.</returns>
-public async Task<byte?> MifareUltralightEV1_CheckTearingEventAsync(byte counterAddr)
-{
-    var parser = await CallFunctionAsync(new byte[] { API_MIFAREULTRALIGHT, 11, counterAddr });
-    var success = parser.ParseBool();
-    if (success)
-    {
-        return parser.ParseByte();
-    }
-    return null;
-}
-
-        #endregion
-
-        #region API_MIFAREDESFIRE / Mifare Desfire Functions
-
-        public static readonly byte API_MIFAREDESFIRE = 15;
-
-        private static readonly byte CRYPTO_ENV = 0;
-        private static readonly byte DESFIRE_AUTHMODE_COMPATIBLE = 0;
-        private static readonly byte DESFIRE_AUTHMODE_EV1 = 1;
-        private static readonly byte DESFIRE_KEYLENGTH = 0x10;
-        private static readonly byte DESFIRE_MAX_FILEIDS = 0xFF;
-
-        private static readonly byte MIFARE_DESFIRE_GETAPPIDS = 0;
-        private static readonly byte MIFARE_DESFIRE_CREATEAPP = 1;
-        private static readonly byte MIFARE_DESFIRE_DELETEAPP = 2;
-        private static readonly byte MIFARE_DESFIRE_SELECTAPP = 3;
-        private static readonly byte MIFARE_DESFIRE_AUTH = 4;
-        private static readonly byte MIFARE_DESFIRE_GETKEYSETTINGS = 5;
-        private static readonly byte MIFARE_DESFIRE_GETFILEIDS = 6;
-        private static readonly byte MIFARE_DESFIRE_GETFILESETTINGS = 7;
-        private static readonly byte MIFARE_DESFIRE_READDATA = 8;
-        private static readonly byte MIFARE_DESFIRE_WRITEDATA = 9;
-        private static readonly byte MIFARE_DESFIRE_GETVALUE = 10;
-        private static readonly byte MIFARE_DESFIRE_CREDIT = 11;
-        private static readonly byte MIFARE_DESFIRE_DEBIT = 12;
-        private static readonly byte MIFARE_DESFIRE_LIMITEDCREDIT = 13;
-        private static readonly byte MIFARE_DESFIRE_GETFREEMEMORY = 14;
-        private static readonly byte MIFARE_DESFIRE_FORMATTAG = 15;
-        private static readonly byte MIFARE_DESFIRE_CREATE_STDDATAFILE = 16;
-        private static readonly byte MIFARE_DESFIRE_CREATE_VALUEFILE = 17;
-        private static readonly byte MIFARE_DESFIRE_GETVERSION = 18;
-        private static readonly byte MIFARE_DESFIRE_DELETEFILE = 19;
-        private static readonly byte MIFARE_DESFIRE_COMMITTRANSACTION = 20;
-        private static readonly byte MIFARE_DESFIRE_ABORTTRANSACTION = 21;
-        private static readonly byte MIFARE_DESFIRE_GETUID = 22;
-        private static readonly byte MIFARE_DESFIRE_GETKEYVERSION = 23;
-        private static readonly byte MIFARE_DESFIRE_CHANGEKEYSETTINGS = 24;
-        private static readonly byte MIFARE_DESFIRE_CHANGEKEY = 25;
-        private static readonly byte MIFARE_DESFIRE_CHANGEFILESETTINGS = 26;
-        private static readonly byte MIFARE_DESFIRE_DISABLEFORMATCARD = 27;
-        private static readonly byte MIFARE_DESFIRE_ENABLERANDOMID = 28;
-        private static readonly byte MIFARE_DESFIRE_SETDEFAULTKEY = 29;
-        private static readonly byte MIFARE_DESFIRE_SETATS = 30;
-        private static readonly byte MIFARE_DESFIRE_CREATERECORDFILE = 31;
-        private static readonly byte MIFARE_DESFIRE_READRECORDS = 32;
-        private static readonly byte MIFARE_DESFIRE_WRITERECORD = 33;
-        private static readonly byte MIFARE_DESFIRE_CLEARRECORDFILE = 34;
-
-        /// <summary>
-        /// Retrieve the Available Application IDs after selecing PICC (App 0), Authentication is needed - depending on the security config
-        /// </summary>
-        /// <returns>a uint32[] of the available appids with 4bytes each, null if no apps are available or on error</returns>
-        public async Task<UInt32[]> MifareDesfire_GetAppIDsAsync()
-        {
-            return await MifareDesfire_GetAppIDsAsync(28);
-        }
-
-        /// <summary>
-        /// Retrieve the Available Application IDs after selecing PICC (App 0), Authentication is needed - depending on the security config
-        /// </summary>
-        /// <param name="maxAppIDCnt"></param>
-        /// <returns>a uint32[] of the available appids with 4bytes each, null if no apps are available or on error</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<UInt32[]> MifareDesfire_GetAppIDsAsync(byte maxAppIDCnt)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_GETAPPIDS , CRYPTO_ENV , maxAppIDCnt};
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                var appIDCnt = parser.ParseByte();
-
-                var appids = new UInt32[appIDCnt];
-
-                for (var i = 0; i < appIDCnt; i++)
-                {
-                    appids[i] = parser.ParseUInt32();
-                }
-
-                return appids;
-            }
-            else
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Creates a new Application
-        /// </summary>
-        /// <param name="_keySettingsTarget">byte: KS_CHANGE_KEY_WITH_MK = 0, KS_ALLOW_CHANGE_MK = 1, KS_FREE_LISTING_WITHOUT_MK = 2, KS_FREE_CREATE_DELETE_WITHOUT_MK = 4, KS_CONFIGURATION_CHANGEABLE = 8, KS_DEFAULT = 11, KS_CHANGE_KEY_WITH_TARGETED_KEYNO = 224, KS_CHANGE_KEY_FROZEN = 240</param>
-        /// <param name="_keyTypeTargetApplication">byte: 0 = 3DES, 1 = 3K3DES, 2 = AES</param>
-        /// <param name="_maxNbKeys">int max. number of keys</param>
-        /// <param name="_appID">int application id</param>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_CreateApplicationAsync(DESFireAppAccessRights keySettingsTarget, DESFireKeyType keyTypeTargetApplication, int maxNbKeys, int appID)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_CREATEAPP, CRYPTO_ENV };
-            bytes.AddUInt32((UInt32)appID);
-            bytes.Add((byte)keySettingsTarget);
-            bytes.AddUInt32((UInt32)maxNbKeys);
-            bytes.AddUInt32((UInt32)keyTypeTargetApplication);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Select a Desfire Application
-        /// </summary>
-        /// <param name="appID"></param>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_DeleteApplicationAsync(uint appID)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_DELETEAPP, CRYPTO_ENV };
-            bytes.AddUInt32((UInt32)appID);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Select a Mifare Desfire Application
-        /// </summary>
-        /// <param name="appID">The Application ID to select</param>
-        /// <returns>true if Application could be selected, false otherwise</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_SelectApplicationAsync(uint appID)
-        {          
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_SELECTAPP, CRYPTO_ENV };
-            bytes.AddUInt32((UInt32)appID);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Authenticate to a previously selected desfire application
-        /// </summary>
-        /// <param name="key">string: a 16 bytes key e.g. 00000000000000000000000000000000</param>
-        /// <param name="keyNo">byte: the keyNo to use</param>
-        /// <param name="keyType">byte: 0 = 3DES, 1 = 3K3DES, 2 = AES</param>
-        /// <param name="authMode">byte: 1 = EV1 Mode, 0 = EV0 Mode</param>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_AuthenticateAsync(string key, byte keyNo, byte keyType, byte authMode)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_AUTH, CRYPTO_ENV , keyNo, DESFIRE_KEYLENGTH};
-            bytes.AddRange(ByteArrayConverter.GetBytesFrom(key));
-            bytes.Add(keyType);
-            bytes.Add(authMode);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Get the KeySettings (Properties: KeySettings, NumberOfKeys, KeyType) of the selected Application. Authentication is needed - depending on the security config
-        /// </summary>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<DESFireKeySettings> MifareDesfire_GetKeySettingsAsync()
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_GETKEYSETTINGS, CRYPTO_ENV};
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                var keySettings = new DESFireKeySettings();
-
-                keySettings.AccessRights = (DESFireAppAccessRights)parser.ParseByte();
-                keySettings.NumberOfKeys = parser.ParseUInt32();
-                keySettings.KeyType = (DESFireKeyType)parser.ParseUInt32();
-
-                return keySettings;
-            }
-            else
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Retrieve the available file IDs after selecing app and authenticating to app
-        /// </summary>
-        /// <returns>byte[] array of available file ids</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<byte[]> MifareDesfire_GetFileIDsAsync()
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_GETFILEIDS, CRYPTO_ENV , DESFIRE_MAX_FILEIDS};
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                var filesCount = parser.ParseByte();
-                var fids = new byte[filesCount];
-
-                for (var i = 0; i < filesCount; i++)
-                {
-                    fids[i] = parser.ParseByte();
-                }
-
-                return fids;
-            }
-            else
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Get the filesettings of a fileid
-        /// </summary>
-        /// <param name="fileNo">ID of the desired file</param>
-        /// <returns><see cref="DESFireFileSettings"/></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<DESFireFileSettings> MifareDesfire_GetFileSettingsAsync(byte fileNo)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_GETFILESETTINGS, CRYPTO_ENV, fileNo };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                var fileSettings = new DESFireFileSettings();
-
-                fileSettings.FileType = (DESFireFileType)parser.ParseByte();
-                fileSettings.ComSett = parser.ParseByte();
-
-                var ar = parser.ParseUInt16();
-
-                fileSettings.accessRights.ReadKeyNo = (byte)(ar & 0x000F);
-                fileSettings.accessRights.WriteKeyNo = (byte)((ar & 0x00F0) >> 4);
-                fileSettings.accessRights.ReadWriteKeyNo = (byte)((ar & 0x0F00) >> 8);
-                fileSettings.accessRights.ChangeKeyNo = (byte)((ar & 0xF000) >> 12);
-
-                switch (fileSettings.FileType)
-                {
-                    case DESFireFileType.DF_FT_STDDATAFILE:
-                    case DESFireFileType.DF_FT_BACKUPDATAFILE:
-                        fileSettings.DataFileSetting = fileSettings.DataFileSetting ?? new DataFileSetting();
-                        fileSettings.DataFileSetting.FileSize = parser.ParseUInt32();
-                        break;
-
-                    case DESFireFileType.DF_FT_VALUEFILE:
-                        fileSettings.ValueFileSetting.LowerLimit = parser.ParseUInt32();
-                        fileSettings.ValueFileSetting.UpperLimit = parser.ParseUInt32();
-                        fileSettings.ValueFileSetting.LimitedCreditValue = parser.ParseUInt32();
-                        fileSettings.ValueFileSetting.LimitedCreditEnabled = parser.ParseByte();
-                        fileSettings.ValueFileSetting.FreeGetValue = parser.ParseByte();
-                        fileSettings.ValueFileSetting.RFU = parser.ParseByte();
-                        break;
-
-                    case DESFireFileType.DF_FT_CYCLICRECORDFILE:
-                    case DESFireFileType.DF_FT_LINEARRECORDFILE:
-                        fileSettings.RecordFileSetting = fileSettings.RecordFileSetting ?? new RecordFileSetting();
-                        fileSettings.RecordFileSetting.RecordSize = parser.ParseUInt32();
-                        fileSettings.RecordFileSetting.MaxNumberOfRecords = parser.ParseUInt32();
-                        fileSettings.RecordFileSetting.CurrentNumberOfRecords = parser.ParseUInt32();
-                        break;
-
-                    default:
-
-                        break;
-                }
-
-                return fileSettings;
-            }
-            else
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Read out Data on a Desfire
-        /// </summary>
-        /// <param name="fileNo">byte: filenumber: 0x00 - 0x14</param>
-        /// <param name="length">int: filesize to read</param>
-        /// <param name="comSet">byte: 0 = Plain, 1 = CMAC, 2 = Encrypted</param>
-        /// <returns>byte[] of data</returns>
-        /// <exception cref="ReaderException"></exception>
-        public Task<byte[]> MifareDesfire_ReadDataAsync(byte fileNo, int length, EncryptionMode mode)
-        {
-            // Simple Protocol: DESFire_ReadData [0F08][Byte: CryptoEnv][Byte: FileNo][UInt16: Offset][Byte: Length][Byte: CommSet]
-            // Length is a single byte, so reads > 255 bytes must be segmented.
-            return MifareDesfire_ReadDataAsync(fileNo, 0, length, mode);
-        }
-
-        /// <summary>
-        /// Read out Data on a DESFire (Simple Protocol: DESFire_ReadData / 0x0F08)
-        /// </summary>
-        /// <param name="fileNo">byte: filenumber</param>
-        /// <param name="offset">UInt16: start offset</param>
-        /// <param name="length">int: number of bytes to read</param>
-        /// <param name="mode">byte: 0 = Plain, 1 = CMAC, 2 = Encrypted</param>
-        /// <returns>byte[] of data</returns>
-        /// <remarks>
-        /// Simple Protocol limits Offset to UInt16 and Length to one byte (max 255). This method transparently
-        /// segments reads larger than 255 bytes by issuing multiple DESFire_ReadData calls with increasing offsets.
-        /// </remarks>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<byte[]> MifareDesfire_ReadDataAsync(byte fileNo, ushort offset, int length, EncryptionMode mode)
-        {
-            if (length < 0) throw new ArgumentOutOfRangeException(nameof(length));
-            if (length == 0) return new byte[0];
-
-            // Offset is UInt16 in Simple Protocol, therefore offset + length must stay within 0..65535.
-            if ((uint)offset + (uint)length > 0x10000)
-                throw new ArgumentOutOfRangeException(nameof(length), "Simple Protocol uses a 16-bit offset; offset + length must be <= 65536.");
-
-            var data = new byte[length];
-            var remaining = length;
-            var dstIndex = 0;
-            var currentOffset = offset;
-
-            while (remaining > 0)
-            {
-                var chunkLen = (byte)Math.Min(0xFF, remaining);
-
-                List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_READDATA, CRYPTO_ENV, fileNo };
-                bytes.AddUInt16(currentOffset);
-                bytes.Add(chunkLen);
-                bytes.Add((byte)mode);
-
-                var parser = await CallFunctionAsync(bytes.ToArray());
-                var success = parser.ParseBool();
-
-                if (!success)
-                    throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-
-                var chunk = parser.ParseVarByteArray();
-                if (chunk.Length != chunkLen)
-                    throw new ReaderException("Unexpected DESFire_ReadData length. Expected " + chunkLen + " bytes, got " + chunk.Length + " bytes.", null);
-
-                Array.Copy(chunk, 0, data, dstIndex, chunkLen);
-
-                dstIndex += chunkLen;
-                remaining -= chunkLen;
-                currentOffset = (ushort)(currentOffset + chunkLen);
-            }
-
-            return data;
-        }
-
-        /// <summary>
-        /// Write Data to a Desfire File
-        /// </summary>
-        /// <param name="fileNo">The file number to read</param>
-        /// <param name="data"></param>
-        /// <param name="mode"><see cref="EncryptionMode"/></param>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public Task MifareDesfire_WriteDataAsync(byte fileNo, byte[] data, EncryptionMode mode)
-        {
-            // Simple Protocol: DESFire_WriteData [0F09][Byte: CryptoEnv][Byte: FileNo][UInt16: Offset][Byte Array(Var): Data][Byte: CommSet]
-            // The 'Data' field is length-prefixed (one byte), so writes > 255 bytes must be segmented.
-            return MifareDesfire_WriteDataAsync(fileNo, 0, data, mode);
-        }
-
-        /// <summary>
-        /// Write Data to a DESFire (Simple Protocol: DESFire_WriteData / 0x0F09)
-        /// </summary>
-        /// <param name="fileNo">byte: filenumber</param>
-        /// <param name="offset">UInt16: start offset</param>
-        /// <param name="data">byte[]: data to write</param>
-        /// <param name="mode">byte: 0 = Plain, 1 = CMAC, 3 = Fully Encrypted</param>
-        /// <remarks>
-        /// In Simple Protocol the data payload is a Byte Array(Var), i.e. it starts with a single length byte.
-        /// Therefore each command can transport at most 255 data bytes. This method transparently segments larger
-        /// payloads by issuing multiple DESFire_WriteData calls with increasing offsets.
-        /// </remarks>
-        public async Task MifareDesfire_WriteDataAsync(byte fileNo, ushort offset, byte[] data, EncryptionMode mode)
-        {
-            if (data == null) throw new ArgumentNullException(nameof(data));
-            if (data.Length == 0) return;
-
-            // Offset is UInt16 in Simple Protocol, therefore offset + data.Length must stay within 0..65535.
-            if ((uint)offset + (uint)data.Length > 0x10000)
-                throw new ArgumentOutOfRangeException(nameof(data), "Simple Protocol uses a 16-bit offset; offset + data.Length must be <= 65536.");
-
-            var remaining = data.Length;
-            var srcIndex = 0;
-            var currentOffset = offset;
-
-            while (remaining > 0)
-            {
-                var chunkLen = Math.Min(0xFF, remaining);
-
-                // Build: [API][FUNC][CryptoEnv][FileNo][UInt16 Offset][VarLen][Data...][CommSet]
-                List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_WRITEDATA, CRYPTO_ENV, fileNo };
-                bytes.AddUInt16(currentOffset);
-                bytes.Add((byte)chunkLen); // Var array length byte
-
-                var chunk = new byte[chunkLen];
-                Array.Copy(data, srcIndex, chunk, 0, chunkLen);
-                bytes.AddRange(chunk);
-
-                bytes.Add((byte)mode);
-
-                var parser = await CallFunctionAsync(bytes.ToArray());
-                var success = parser.ParseBool();
-
-                if (!success)
-                    throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-
-                srcIndex += chunkLen;
-                remaining -= chunkLen;
-                currentOffset = (ushort)(currentOffset + chunkLen);
-            }
-        }
-
-                /// <summary>
-        /// Get the current value of a DESFire value file.
-        /// </summary>
-        /// <param name="fileNo">File number (value file)</param>
-        /// <param name="mode">Communication setting (Plain/CMAC/Encrypted)</param>
-        /// <returns>The current value.</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<UInt32> MifareDesfire_GetValueAsync(byte fileNo, EncryptionMode mode)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_GETVALUE, CRYPTO_ENV, fileNo, (byte)mode };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                return parser.ParseUInt32();
-            }
-
-            throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-        }
-
-        /// <summary>
-        /// Credit (add) a value to a DESFire value file (transactional).
-        /// </summary>
-        /// <param name="fileNo">File number (value file)</param>
-        /// <param name="value">Value to credit</param>
-        /// <param name="mode">Communication setting (Plain/CMAC/Encrypted)</param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_CreditAsync(byte fileNo, UInt32 value, EncryptionMode mode)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_CREDIT, CRYPTO_ENV, fileNo };
-
-            bytes.AddUInt32(value);
-            bytes.Add((byte)mode);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Debit (subtract) a value from a DESFire value file (transactional).
-        /// </summary>
-        /// <param name="fileNo">File number (value file)</param>
-        /// <param name="value">Value to debit</param>
-        /// <param name="mode">Communication setting (Plain/CMAC/Encrypted)</param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_DebitAsync(byte fileNo, UInt32 value, EncryptionMode mode)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_DEBIT, CRYPTO_ENV, fileNo };
-
-            bytes.AddUInt32(value);
-            bytes.Add((byte)mode);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Limited credit a value to a DESFire value file (transactional).
-        /// </summary>
-        /// <param name="fileNo">File number (value file)</param>
-        /// <param name="value">Value to credit</param>
-        /// <param name="mode">Communication setting (Plain/CMAC/Encrypted)</param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_LimitedCreditAsync(byte fileNo, UInt32 value, EncryptionMode mode)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_LIMITEDCREDIT, CRYPTO_ENV, fileNo };
-
-            bytes.AddUInt32(value);
-            bytes.Add((byte)mode);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Get the free Memory of a desfire. 
-        /// </summary>
-        /// <returns>a uint32 of the available memory if supported</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<UInt16> MifareDesfire_GetFreeMemoryAsync()
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_GETFREEMEMORY, CRYPTO_ENV };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                return parser.ParseUInt16();
-            }
-            else
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Format a Chip
-        /// </summary>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_FormatTagAsync()
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_FORMATTAG, CRYPTO_ENV };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Create a std data file on a desfire
-        /// </summary>
-        /// <param name="fileNo"></param>
-        /// <param name="fileType"><see cref="DESFireFileType"/></param>
-        /// <param name="mode"><see cref="EncryptionMode"/></param>
-        /// <param name="accessRights"><see cref="DESFireFileAccessRights"/></param>
-        /// <param name="fileSize"></param>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_CreateStdDataFileAsync(byte fileNo, DESFireFileType fileType, EncryptionMode mode, DESFireFileAccessRights accessRights, UInt32 fileSize)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_CREATE_STDDATAFILE, CRYPTO_ENV, fileNo, (byte)fileType, (byte)mode };
-
-            UInt16 fileAccessRights = 0;
-
-            fileAccessRights |= accessRights.ReadKeyNo;
-            fileAccessRights |= (byte)(accessRights.WriteKeyNo << 4);
-            fileAccessRights |= (byte)(accessRights.ReadWriteKeyNo << 8);
-            fileAccessRights |= (byte)(accessRights.ChangeKeyNo << 12);
-
-            bytes.AddUInt16(fileAccessRights);
-            bytes.AddUInt32(fileSize);
-            bytes.AddRange(new byte[] { 0,0,0,0, 0,0,0,0, 0,0,0,0 });
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-        /// <summary>
-        /// Create a value file on a DESFire.
-        /// </summary>
-        /// <param name="fileNo"></param>
-        /// <param name="fileType"><see cref="DESFireFileType"/></param>
-        /// <param name="mode"><see cref="EncryptionMode"/></param>
-        /// <param name="accessRights"><see cref="DESFireFileAccessRights"/></param>
-        /// <param name="lowerLimit"></param>
-        /// <param name="upperLimit"></param>
-        /// <param name="limitedCreditValue"></param>
-        /// <param name="freeGetValue">If true, GetValue is allowed without authentication (bit 0).</param>
-        /// <param name="limitedCreditEnabled">If true, LimitedCredit is enabled (bit 1).</param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_CreateValueFileAsync(
-            byte fileNo,
-            DESFireFileType fileType,
-            EncryptionMode mode,
-            DESFireFileAccessRights accessRights,
-            UInt32 lowerLimit,
-            UInt32 upperLimit,
-            UInt32 limitedCreditValue,
-            bool freeGetValue,
-            bool limitedCreditEnabled)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_CREATE_VALUEFILE, CRYPTO_ENV, fileNo, (byte)fileType, (byte)mode };
-
-            UInt16 fileAccessRights = 0;
-
-            fileAccessRights |= accessRights.ReadKeyNo;
-            fileAccessRights |= (UInt16)(accessRights.WriteKeyNo << 4);
-            fileAccessRights |= (UInt16)(accessRights.ReadWriteKeyNo << 8);
-            fileAccessRights |= (UInt16)(accessRights.ChangeKeyNo << 12);
-
-            bytes.AddUInt16(fileAccessRights);
-            bytes.AddUInt32(lowerLimit);
-            bytes.AddUInt32(upperLimit);
-            bytes.AddUInt32(limitedCreditValue);
-
-            UInt32 flags = 0;
-            if (freeGetValue) flags |= 0x01;
-            if (limitedCreditEnabled) flags |= 0x02;
-            bytes.AddUInt32(flags);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-        /// <summary>
-        /// Get version of a desfire.
-        /// </summary>
-        /// <returns></returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<byte[]> MifareDesfire_GetVersionAsync()
-        {
-            {
-                List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_GETVERSION, CRYPTO_ENV };
-
-                var parser = await CallFunctionAsync(bytes.ToArray());
-                var success = parser.ParseBool();
-
-                if (success)
-                {
-                    return parser.ParseVarByteArray();
-                }
-                else
-                {
-                    throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.NotSupported), null);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Delete a file in a desfire app
-        /// </summary>
-        /// <param name="fileNo">byte: Filenumber to delete</param>
-        /// <returns>true if the Operation was successful, false otherwise</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_DeleteFileAsync(byte fileNo)
-        {
-            {
-                List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_DELETEFILE, CRYPTO_ENV, fileNo };
-
-                var parser = await CallFunctionAsync(bytes.ToArray());
-                var success = parser.ParseBool();
-
-                if (!success)
-                {
-                    throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Commit a DESFire transaction (e.g., after Credit/Debit/LimitedCredit).
-        /// </summary>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_CommitTransactionAsync()
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_COMMITTRANSACTION, CRYPTO_ENV };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Abort a DESFire transaction.
-        /// </summary>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_AbortTransactionAsync()
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_ABORTTRANSACTION, CRYPTO_ENV };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Get the UID from the currently selected DESFire card.
-        /// </summary>
-        /// <param name="bufferSize">Maximum UID bytes to return (default: 0xFF)</param>
-        /// <returns>UID byte array</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<byte[]> MifareDesfire_GetUidAsync(byte bufferSize = 0xFF)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_GETUID, CRYPTO_ENV, bufferSize };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                return parser.ParseVarByteArray();
-            }
-
-            throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-        }
-
-        /// <summary>
-        /// Get the key version of a DESFire key.
-        /// </summary>
-        /// <param name="keyNo">Key number</param>
-        /// <returns>Key version</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<byte> MifareDesfire_GetKeyVersionAsync(byte keyNo)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_GETKEYVERSION, CRYPTO_ENV, keyNo };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                return parser.ParseByte();
-            }
-
-            throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-        }
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="keySettings"></param>
-        /// <param name="numberOfKeys"></param>
-        /// <param name="keyType"></param>
-        /// <returns></returns>
-        public async Task MifareDesfire_ChangeKeySettingsAsync(DESFireAppAccessRights keySettings, UInt32 numberOfKeys, DESFireKeyType keyType)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_CHANGEKEYSETTINGS, CRYPTO_ENV};
-
-            bytes.Add((byte)keySettings);
-            bytes.AddUInt32(numberOfKeys);
-            bytes.AddUInt32((byte)keyType);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Changes a Key
-        /// </summary>
-        /// <param name="oldKey"></param>
-        /// <param name="newKey"></param>
-        /// <param name="keyVersion"></param>
-        /// <param name="accessRights"></param>
-        /// <param name="keyNo"></param>
-        /// <param name="numberOfKeys"></param>
-        /// <param name="keyType">The Type of the new Key</param>
-        /// <returns></returns>
-        public async Task MifareDesfire_ChangeKeyAsync(string oldKey, string newKey, byte keyVersion, byte accessRights, byte keyNo, UInt32 numberOfKeys, DESFireKeyType keyType)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_CHANGEKEY, CRYPTO_ENV, keyNo };
-            bytes.Add(DESFIRE_KEYLENGTH);
-            bytes.AddRange(ByteArrayConverter.GetBytesFrom(oldKey));
-            bytes.Add(DESFIRE_KEYLENGTH);
-            bytes.AddRange(ByteArrayConverter.GetBytesFrom(newKey));
-            bytes.Add(keyVersion);
-            bytes.Add(accessRights);
-            bytes.AddUInt32(numberOfKeys);
-            bytes.AddUInt32((byte)keyType);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Change a DESFire file's communication settings and access rights.
-        /// </summary>
-        /// <param name="fileNo"></param>
-        /// <param name="newCommSet"><see cref="EncryptionMode"/></param>
-        /// <param name="oldAccessRights"><see cref="DESFireFileAccessRights"/></param>
-        /// <param name="newAccessRights"><see cref="DESFireFileAccessRights"/></param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_ChangeFileSettingsAsync(byte fileNo, EncryptionMode newCommSet, DESFireFileAccessRights oldAccessRights, DESFireFileAccessRights newAccessRights)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_CHANGEFILESETTINGS, CRYPTO_ENV, fileNo, (byte)newCommSet };
-
-            UInt16 oldAr = 0;
-            oldAr |= oldAccessRights.ReadKeyNo;
-            oldAr |= (UInt16)(oldAccessRights.WriteKeyNo << 4);
-            oldAr |= (UInt16)(oldAccessRights.ReadWriteKeyNo << 8);
-            oldAr |= (UInt16)(oldAccessRights.ChangeKeyNo << 12);
-
-            UInt16 newAr = 0;
-            newAr |= newAccessRights.ReadKeyNo;
-            newAr |= (UInt16)(newAccessRights.WriteKeyNo << 4);
-            newAr |= (UInt16)(newAccessRights.ReadWriteKeyNo << 8);
-            newAr |= (UInt16)(newAccessRights.ChangeKeyNo << 12);
-
-            bytes.AddUInt16(oldAr);
-            bytes.AddUInt16(newAr);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Disable FormatCard for the currently selected DESFire card.
-        /// </summary>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_DisableFormatCardAsync()
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_DISABLEFORMATCARD, CRYPTO_ENV };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Enable random UID (RandomID) for the currently selected DESFire card.
-        /// </summary>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_EnableRandomIdAsync()
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_ENABLERANDOMID, CRYPTO_ENV };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Set the default DESFire key (master key) and its version.
-        /// </summary>
-        /// <param name="key">Key material (variable length)</param>
-        /// <param name="keyVersion">Key version byte</param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_SetDefaultKeyAsync(byte[] key, byte keyVersion)
-        {
-            if (key == null) throw new ArgumentNullException(nameof(key));
-            if (key.Length > 0xFF) throw new ArgumentOutOfRangeException(nameof(key), "Key is too long for a Var Byte Array (max 255).");
-
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_SETDEFAULTKEY, CRYPTO_ENV, (byte)key.Length };
-
-            bytes.AddRange(key);
-            bytes.Add(keyVersion);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Set the ATS (Answer To Select) for the currently selected DESFire card.
-        /// </summary>
-        /// <param name="ats">ATS bytes</param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_SetAtsAsync(byte[] ats)
-        {
-            if (ats == null) throw new ArgumentNullException(nameof(ats));
-            if (ats.Length > 0xFF) throw new ArgumentOutOfRangeException(nameof(ats), "ATS is too long for a Var Byte Array (max 255).");
-
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_SETATS, CRYPTO_ENV, (byte)ats.Length };
-
-            bytes.AddRange(ats);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Create a record file on a DESFire.
-        /// </summary>
-        /// <param name="fileNo"></param>
-        /// <param name="fileType"><see cref="DESFireFileType"/> (Linear or Cyclic record file)</param>
-        /// <param name="mode"><see cref="EncryptionMode"/></param>
-        /// <param name="accessRights"><see cref="DESFireFileAccessRights"/></param>
-        /// <param name="recordSize"></param>
-        /// <param name="maxNumberOfRecords"></param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_CreateRecordFileAsync(byte fileNo, DESFireFileType fileType, EncryptionMode mode, DESFireFileAccessRights accessRights, UInt32 recordSize, UInt32 maxNumberOfRecords)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_CREATERECORDFILE, CRYPTO_ENV, fileNo, (byte)fileType, (byte)mode };
-
-            UInt16 fileAccessRights = 0;
-
-            fileAccessRights |= accessRights.ReadKeyNo;
-            fileAccessRights |= (UInt16)(accessRights.WriteKeyNo << 4);
-            fileAccessRights |= (UInt16)(accessRights.ReadWriteKeyNo << 8);
-            fileAccessRights |= (UInt16)(accessRights.ChangeKeyNo << 12);
-
-            bytes.AddUInt16(fileAccessRights);
-            bytes.AddUInt32(recordSize);
-            bytes.AddUInt32(maxNumberOfRecords);
-
-            // appending 0's (8 bytes) per Simple Protocol
-            bytes.AddRange(new byte[] { 0, 0, 0, 0, 0, 0, 0, 0 });
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Read records from a DESFire record file.
-        /// </summary>
-        /// <param name="fileNo"></param>
-        /// <param name="offset"></param>
-        /// <param name="numberOfRecords"></param>
-        /// <param name="recordSize"></param>
-        /// <param name="mode"><see cref="EncryptionMode"/></param>
-        /// <returns>Record data</returns>
-        /// <exception cref="ReaderException"></exception>
-        public async Task<byte[]> MifareDesfire_ReadRecordsAsync(byte fileNo, UInt16 offset, byte numberOfRecords, byte recordSize, EncryptionMode mode)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_READRECORDS, CRYPTO_ENV, fileNo };
-
-            bytes.AddUInt16(offset);
-            bytes.Add(numberOfRecords);
-            bytes.Add(recordSize);
-            bytes.Add((byte)mode);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (success)
-            {
-                return parser.ParseVarByteArray();
-            }
-
-            throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-        }
-
-        /// <summary>
-        /// Write a record to a DESFire record file.
-        /// </summary>
-        /// <param name="fileNo"></param>
-        /// <param name="offset"></param>
-        /// <param name="data"></param>
-        /// <param name="mode"><see cref="EncryptionMode"/></param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_WriteRecordAsync(byte fileNo, UInt16 offset, byte[] data, EncryptionMode mode)
-        {
-            if (data == null) throw new ArgumentNullException(nameof(data));
-            if (data.Length > 0xFF) throw new ArgumentOutOfRangeException(nameof(data), "Data is too long for a Var Byte Array (max 255).");
-
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_WRITERECORD, CRYPTO_ENV, fileNo };
-
-            bytes.AddUInt16(offset);
-            bytes.Add((byte)data.Length);
-            bytes.AddRange(data);
-            bytes.Add((byte)mode);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-
-        /// <summary>
-        /// Clear a DESFire record file.
-        /// </summary>
-        /// <param name="fileNo"></param>
-        /// <exception cref="ReaderException"></exception>
-        public async Task MifareDesfire_ClearRecordFileAsync(byte fileNo)
-        {
-            List<byte> bytes = new List<byte> { API_MIFAREDESFIRE, MIFARE_DESFIRE_CLEARRECORDFILE, CRYPTO_ENV, fileNo };
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-
-            if (!success)
-            {
-                throw new ReaderException("Call was not successfull, error " + Enum.GetName(typeof(ReaderError), ReaderError.AccessDenied), null);
-            }
-        }
-        #endregion
-
-        #region API_ISO14443 / ISO14443 Transparent Transponder Access Functions
-
-        public static readonly byte API_ISO14443 = 18;
-
-        /// <summary>
-        /// This function delivers the ATS (Answer To Select) of a ISO14443A layer 4 transponder.
-        /// </summary>
-        /// <returns>The ATS if one is found, otherwise null.</returns>
-        public async Task<byte[]> ISO14443A_GetAtsAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_ISO14443, 0, /* maxBytes */ byte.MaxValue });
-            var success = parser.ParseBool();
-            if (success)
-            {
-                var result = parser.ParseVarByteArray();
-                return result;
-            }
-            return null;
-        }
-
-        /// <summary>
-        ///     This function delivers the ATQB (Answer To Request TypeB) of the last detected ISO14443B compliant transponder.<br/>
-        ///     Note: This function cannot be called on TWN4 MultiTech Legic.
-        /// </summary>
-        /// <returns>The ATQB if one is found, otherwise null.</returns>
-        public async Task<byte[]> ISO14443B_GetAtqbAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_ISO14443, 1, /* maxBytes */ byte.MaxValue });
-            var success = parser.ParseBool();
-            if (success)
-            {
-                var result = parser.ParseVarByteArray();
-                return result;
-            }
-            return null;
-        }
-
-        /// <summary>
-        ///     This function can be used to probe if a ISO14443-4 transponder is still in reading range. The internal state
-        ///     of the transponder remains unchanged. <br/>
-        ///     Note: This function cannot be called on TWN4 MultiTech Legic.
-        /// </summary>
-        /// <returns></returns>
-        public async Task<bool> ISO14443_4_CheckPresenceAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_ISO14443, 2 });
-            var result = parser.ParseBool();
-            return result;
-        }
-
-        /// <summary>
-        ///     This function can be used for transparent exchange of data between reader and ISO14443-4 transponders.
-        ///     All framing of layer 4 subset is already done by the reader, so only the payload needs to be passed
-        ///     to the function.
-        /// </summary>
-        /// <param name="TX">Data that shall be transmitted to the transponder.</param>
-        /// <returns>The response of the transponder.</returns>
-        public async Task<byte[]> ISO14443_4_TdxAsync(byte[] TX)
-        {
-            List<byte> bytes = new List<byte> { API_ISO14443, 3 };
-            bytes.Add((byte)TX.Length);
-            bytes.AddRange(TX);
-            bytes.Add(byte.MaxValue); // MaxRXByteCnt
-            
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-            if (success)
-            {
-                var result = parser.ParseVarByteArray();
-                return result;
-            }
-            return null;
-        }
-
-        /// <summary>
-        ///     This function delivers the ATQA (Answer To Request TypeA) of the last detected ISO14443A compliant transponder.
-        ///     The ATQA consists of two bytes, parsed in LSB-first order.
-        /// </summary>
-        /// <returns>The ATQA if one is found, otherwise null.</returns>
-        public async Task<ushort?> ISO14443A_GetAtqaAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_ISO14443, 4 });
-            var success = parser.ParseBool();
-            if (success)
-            {
-                var result = parser.ParseUInt16();
-                return result;
-            }
-            return null;
-        }
-
-        /// <summary>
-        /// This function delivers the SAK (Select Acknowledge) of the last detected ISO14443A compliant transponder.
-        /// </summary>
-        /// <returns>The SAK if one is found, otherwise null.</returns>
-        public async Task<byte?> ISO14443A_GetSakAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_ISO14443, 5 });
-            var success = parser.ParseBool();
-            if (success)
-            {
-                var result = parser.ParseByte();
-                return result;
-            }
-            return null;
-        }
-
-        /// <summary>
-        ///     This function delivers the transponder’s answer to the ATTRIB command, which is sent automatically
-        ///     during selection process by the reader. <br/>
-        ///     Note: This function cannot be called on TWN4 MultiTech Legic.
-        /// </summary>
-        /// <returns>The response of the transponder.</returns>
-        public async Task<byte[]> ISO14443B_GetAnswerToAttribAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_ISO14443, 6, /* maxBytes */ byte.MaxValue });
-            var success = parser.ParseBool();
-            if (success)
-            {
-                var result = parser.ParseVarByteArray();
-                return result;
-            }
-            return null;
-        }
-
-        /// <summary>
-        ///     This function can be used for transparent exchange of data between reader and ISO14443-3 transponders.
-        ///     The function does not calculate any CRC or other overhead by itself, so if necessary this has to be
-        ///     conducted on host side.
-        /// </summary>
-        /// <param name="TX">Data that shall be transmitted to the transponder.</param>
-        /// <param name="timeout">Response timeout in milliseconds.</param>
-        /// <returns>The response of the transponder.</returns>
-        public async Task<byte[]> ISO14443_3_TdxAsync(byte[] TX, ushort timeout)
-        {
-            List<byte> bytes = new List<byte> { API_ISO14443, 7 };
-            bytes.Add((byte)TX.Length);
-            bytes.AddRange(TX);
-            bytes.Add(byte.MaxValue); // MaxRXByteCnt
-            bytes.AddUInt16(timeout);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-            if (success)
-            {
-                var result = parser.ParseVarByteArray();
-                return result;
-            }
-            return null;
-        }
-
-        /// <summary>
-        /// Use this function to search the RF field for ISO14443A transponders. The result is a list of the UID of the respective transponders.
-        /// </summary>
-        /// <returns>A list containing the UIDs of all transponders.</returns>
-        public async Task<List<byte[]>> ISO14443A_SearchMultiTagAsync()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_ISO14443, 8, /* maxIDBytes */ byte.MaxValue });
-            var tagList = new List<byte[]>();
-
-            var found = parser.ParseBool();
-            if (found)
-            {
-                var count = parser.ParseByte();
-                parser.ParseByte(); // Total number of bytes. We don't need this.
-                for (int i = 0; i < count; i++)
-                {
-                    var tag = parser.ParseVarByteArray();
-                    tagList.Add(tag);
-                }
-            }
-
-            return tagList;
-        }
-
-
-        /// <summary>
-        /// Use this function to select one of the discovered transponders for further operations.
-        /// IMPORTANT: This does not work on Legic capable TWN4 Mutitec Readers. Use SearchTag instead. 
-        /// </summary>
-        /// <param name="uid">Specify the UID of the transponder to be selected.</param>
-        /// <returns>If the operation was successful, the return value is true, otherwise it is false.</returns>
-        public async Task<bool> ISO14443A_SelectTagAsync(byte[] uid)
-        {
-            List<byte> bytes = new List<byte> { API_ISO14443, 9 };
-            bytes.Add((byte)uid.Length);
-            bytes.AddRange(uid);
-
-            var parser = await CallFunctionAsync(bytes.ToArray());
-            var success = parser.ParseBool();
-            return success;
-        }
-
-        // TODO: SYSFUNC(API_ISO14443, 10, bool preISO14443B_GetATR(byte* ATR, int* ATRByteCnt, int MaxATRByteCnt))
-
-        /// <summary>
-        /// Reselect a transponder.
-        /// </summary>
-        /// <returns></returns>
-        /// <remarks>SYSFUNC(API_ISO14443, 11, bool ISO14443A_Reselect(void))<br/>
-        ///     This internal method is not documented in TWN4 API reference.</remarks>
-        public async Task<bool> ISO14443A_Reselect()
-        {
-            var parser = await CallFunctionAsync(new byte[] { API_ISO14443, 11 });
-            var result = parser.ParseBool();
-            return result;
-        }
-
-        #endregion
-
-        #endregion
 
         #region High Level APIs
